Add improved event orchestrator. Extend matcher interface. Drop pointers for matchers. Remove avatar cache in favour of just using browser.

+++ rules/schema.go
@@ -130,7 +130,7 @@ type PlayerListSchema struct {
 
 type PlayerLastSeen struct {
 	PlayerName string `json:"player_name,omitempty"`
-	Time       int    `json:"time,omitempty"`
+	Time       int64  `json:"time,omitempty"`
 }
 
 type PlayerDefinition struct {
